+++ react/ProductSummaryWishlist.tsx
@@ -5,6 +5,7 @@ import { ExtensionPoint, useRuntime, useTreePath } from 'vtex.render-runtime'
 import { useListContext, ListContextProvider } from 'vtex.list-context'
 import { ProductListContext } from 'vtex.product-list-context'
 import { Spinner } from 'vtex.styleguide'
+import { usePixel } from 'vtex.pixel-manager'
 
 import { mapCatalogProductToProductSummary } from './utils/normalize'
 import ProductListEventCaller from './components/ProductListEventCaller'
@@ -51,6 +52,7 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
   const { list } = useListContext() || []
   const { treePath } = useTreePath()
   const { navigate, history } = useRuntime()
+  const { push } = usePixel()
 
   const sessionResponse: any = useSessionResponse()
 
@@ -127,12 +129,23 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
           }
         }
       }
+
+      const handleOnClick = () =>{
+        push({
+          event: 'productClick',
+          list: 'wishlist',
+          product: normalizedProduct,
+          position: index,
+        })
+      }
+
       return (
         <ExtensionPoint
           id="product-summary"
           key={product.id}
           treePath={treePath}
           product={normalizedProduct}
+          actionOnClick={handleOnClick}
         />
       )
     })